--- a/main/src/ca/uwaterloo/flix/api/Flix.scala
+++ b/main/src/ca/uwaterloo/flix/api/Flix.scala
@@ -144,9 +144,6 @@ class Flix {
     // Reflect
     "Reflect.flix" -> LocalResource.get("/src/library/Reflect.flix"),
 
-    // Debug
-    "Debug.flix" -> LocalResource.get("/src/library/Debug.flix"),
-
     // References
     "Ref.flix" -> LocalResource.get("/src/library/Ref.flix"),
   )
@@ -166,6 +163,7 @@ class Flix {
     "Char.flix" -> LocalResource.get("/src/library/Char.flix"),
     "CodePoint.flix" -> LocalResource.get("/src/library/CodePoint.flix"),
     "Console.flix" -> LocalResource.get("/src/library/Console.flix"),
+    "Debug.flix" -> LocalResource.get("/src/library/Debug.flix"),
     "DelayList.flix" -> LocalResource.get("/src/library/DelayList.flix"),
     "DelayMap.flix" -> LocalResource.get("/src/library/DelayMap.flix"),
     "Down.flix" -> LocalResource.get("/src/library/Down.flix"),