--- a/main/src/ca/uwaterloo/flix/api/FlixEvent.scala
+++ b/main/src/ca/uwaterloo/flix/api/FlixEvent.scala
@@ -15,7 +15,7 @@
  */
 package ca.uwaterloo.flix.api
 
-import ca.uwaterloo.flix.language.ast.{ReducedAst, Symbol}
+import ca.uwaterloo.flix.language.ast.{ReducedAst, Symbol, TypedAst}
 import ca.uwaterloo.flix.language.phase.typer.TypeConstraint
 import ca.uwaterloo.flix.language.phase.unification.set.Equation
 
@@ -30,6 +30,11 @@ object FlixEvent {
     * An event that is fired after the tailpos phase.
     */
   case class AfterTailPos(root: ReducedAst.Root) extends FlixEvent
+
+  /**
+    * An event that is fired after the Typer phase.
+    */
+  case class AfterTyper(root: TypedAst.Root) extends FlixEvent
 
   /**
     * An event that is fired when new type constraints are collected for the given def symbol `sym`.