… README.md
… main/src/ca/uwaterloo/flix/api/lsp/provider/completion/OpCompleter.scala
… main/src/ca/uwaterloo/flix/api/lsp/provider/completion/SignatureCompleter.scala
… main/src/ca/uwaterloo/flix/language/ast/ops/TypedAstOps.scala
… main/src/ca/uwaterloo/flix/language/dbg/printer/MonoAstPrinter.scala
… main/src/ca/uwaterloo/flix/language/phase/EnumOptimizer.scala
… main/src/ca/uwaterloo/flix/language/phase/Kinder.scala
… main/src/ca/uwaterloo/flix/language/phase/Monomorpher.scala
… main/src/ca/uwaterloo/flix/language/phase/OccurrenceAnalyzer.scala
… main/src/ca/uwaterloo/flix/language/phase/Simplifier.scala
… main/src/ca/uwaterloo/flix/language/phase/TypeReconstruction.scala
… main/src/ca/uwaterloo/flix/language/phase/extra/CodeHinter.scala
… main/src/ca/uwaterloo/flix/language/phase/optimizer/Inliner.scala
… main/src/ca/uwaterloo/flix/language/phase/optimizer/LambdaDrop.scala
… main/src/ca/uwaterloo/flix/language/phase/optimizer/OccurrenceAnalyzer.scala
… main/src/ca/uwaterloo/flix/language/phase/optimizer/Optimizer.scala
… main/src/ca/uwaterloo/flix/tools/pkg/github/GitHub.scala
… main/src/ca/uwaterloo/flix/util/ClassList.scala
… main/src/ca/uwaterloo/flix/util/LocalResource.scala
… main/src/ca/uwaterloo/flix/util/Options.scala
… main/src/ca/uwaterloo/flix/util/Result.scala
… main/src/ca/uwaterloo/flix/util/Similarity.scala
… main/src/ca/uwaterloo/flix/util/collection/CofiniteIntSet.scala
… main/src/ca/uwaterloo/flix/util/collection/ListMap.scala
… main/src/library/Debug.flix
… main/src/library/Fixpoint/Boxable.flix
… main/src/library/Identity.flix
… main/src/library/SemiGroup.flix
… main/src/library/Time.flix
… main/src/resources/benchmark/Benchmark.Enum.flix