@magnus-madsen 2025-05-09
Did you accidentally merge the wrong branch? This seems to have the JvmReflect stuff in it too.