@JonathanStarup 2025-05-10
> Did you accidentally merge the wrong branch? This seems to have the JvmReflect stuff in it too.

This requires JvmReflection. It's described in the main text, but the change of enum representation exposes a coupling of type match and jvm stuff that doesn't hold here.