@magnus-madsen 2025-07-07
I assume that with the new MonoType and SimpleType we will need a clean design here?