@magnus-madsen 2025-07-07
I don't want a solution that introduces another typematch-like construct.

I just want a simple AST rewrite.