@magnus-madsen 2025-10-04
Closing -- but we are re-implementing this in small pieces.